sh/vim/tmux: Remove traces of base16
Base16 is a very interesting framework for themes, but it has a number
of limitations that make it suboptimal for use in editors, particularly
Vim.

Besides, the project has been trying to renew itself in the absence of
its original author — which is a good initiative — but it might take a
while before the new spec becomes stable.

A good summary:
https://github.com/base16-project/home/issues/52#issuecomment-1177305079