nvim: Do not map <CR> to cmp.confirm in insert mode
This mapping was suggested in the docs. In practice, auto-selecting the
first entry on Enter often results in an undesired entry being inserted
instead of a newline, which is annoying.

The mapping presets in insert mode are actually sane and don't need to
be augmented IMO: confirm with <C-y> and abort with <C-e>.

https://github.com/hrsh7th/nvim-cmp/blob/v0.0.1/lua/cmp/config/mapping.lua#L37-L42