nvim: Fallback method for host clipboard integration
Seems to work reliably for copying and pasting to/from the Windows
clipboard from WSL, but causes permanent garbage in the terminal after
interacting with the clipboard. Invoking PowerShell in interoperability
mode is also extremely slow, so using win32yank.exe (which is natively
supported by Neovim) should be preferred over this.