sh: Skip kubectl autocompletion unless config exists
Rancher Desktop uses a kubectl executable which is actually kuberlr.
kuberlr attempts to use the active kubeconfig context to determine what
kubectl version to use (localhost:8080 if not set). This causes a major
performance hit while loading the zshrc, due to the inevitable timeout
when that cluster isn't reachable.

Ref.
  https://github.com/rancher-sandbox/rancher-desktop/blob/v1.7.0/scripts/dependencies/tools.ts#L120-L124
  https://github.com/flavio/kuberlr