nvim: Revamp statusline
Minimal design without powerline elements. Uses few highlights to be
able to support virtually any colorscheme, and provides these highlights
exclusively through User{N} highlight groups for simplicity.
(:h hl-User1..9)

Elements are now rendered through "%{}" items that expand expressions
returned by Lua functions. These are evaluated in the context of the
window that the statusline belongs to, unlike "%!", which is evaluated
in the context of the current window, and therefore propagates to all
windows. As a result, using a 'laststatus' value other than 3 now works
as expected.
(:h 'statusline')