make: More rigorous naming conventions
Variables representing constants a user might want to customize on the
command line or in the environment are written in all uppercase with
words separated by underscores.

Variables that appear only in the makefile are all lowercase with words
separated by underscores.

User-defined functions in variables and macros use lowercase words
separated by dashes.