make: Avoid unnecessary exec of 'zimfw install'
Only install Zim modules if additional modules are missing. The 'fzf'
module is used as a reference because it isn't part of the default Zim
modules and is included in my dot files regardless of the platform.

The .zimrc file is used as a dependency of the modules goal to ensure
that 'zimfw install' is only executed after the custom dot files have
been made/installed.

The respective files are touched at the end of each goal to compensate
for the regular updates of the ~/.zim directory, which can cause its
dependants to be unnecessarily re-made.