nix: Revert startup key swap on USB keyboard
This reverts commit 203340d1c5c9e7d076e4d2142c102a341fc3a20b.

HIDUTIL isn't ergonomic when it comes to configuring USB devices because
key remappings are lost when the keyboard service is removed[1], which
seems to happen whenever the keyboard is powered off (e.g. the external
display it is connected to is powered off).

The problem doesn't exist when key remappings are configured manually
through the macOS System Settings.

[1]: https://developer.apple.com/library/archive/technotes/tn2450/_index.html