nix: Revert include lldb in Rust devshell
This reverts commit 2b848cbcc475c703134e2ceee1d0254bf5153b16.

Xcode ships with its own 'lldb' executable, which will be executed
through the /usr/bin/lldb stub (calls `xcrun --find lldb` under the
hood) as long as the DEVELOPER_DIR environment variable is not set (cf.
output of `xcode-select -p`). This makes the installation of another
copy of LLDB unnecessary.

There is a catch with Nix devShells: the DEVELOPER_DIR environment
variable gets set to the path of the apple-sdk package in the Nix
store[1] (also visible in the output of `xcode-select -p`). This
prevents the /usr/bin/lldb stub from finding Xcode's 'lldb' executable
("error: tool 'lldb' not found").

With the 'lldb' package from nixpkgs this issue doesn't exist because
'lldb' is in the PATH, but rust-lldb prints "error: unable to locate
debugserver" upon execution of the program through the debugger, unless
LLDB_DEBUGSERVER_PATH is set to the path of the 'debugserver' executable
provided by Xcode[2].

[1]: NixOS/nixpkgs#355486
[2]: NixOS/nixpkgs#252838