sh: Export FZF_DEFAULT_OPTS as a string
An array parameter simplifies adding CLI arguments to the environmental
variable and prevents inserting spaces in place of empty parameters.
However, exported array parameters aren't returned by the getenv(3)
syscall, so fzf did not pick the defaults.

Array parameters can also not be re-exported as a string after they have
been declared as an array ('inconsistent type for assignment'), so a
temporary local parameter is used to concatenate the CLI arguments.