qmk: Revamp Numbers and Symbols layers
Numbers on the top row feel uncomfortable to type with the 'lower' thumb
key pressed. Place them all on the opposite hand like a numpad instead.
Additionally, place symbols frequently typed in relation with numbers on
the remaining keys.

For the same reason, symbols on the top row feel uncomfortable to reach
with the 'upper' thumb key pressed. Spread them across both hands
instead, taking inspiration from sunaku's symbol layer[1].
The most frequently hit symbols are placed on the right hand, in a
Vim-friendly arrangement rather than a mirror of the ANSI layout.
Most symbols on the left hand are duplicated from the base layer for
convenience.

Finally, return to the OLKB style "Tri Layer" system[2] for layer keys,
which was the default on Vial.

[1]: https://sunaku.github.io/moergo-glove80-keyboard.html#symbol-layer
[2]: https://docs.qmk.fm/features/tri_layer