qmk: Free the inner thumb keys
The assignment of frequently accessed functions (Space, Backspace,
Enter, Tab) to inner thumb keys on the Corne is causing me pain after
prolonged periods of work.

Repurpose the outer thumb keys as Top-Hold keys that activate layers on
hold, and move Enter and Tab to a more conventional location on outer
columns.