qmk: Minor changes to Numbers and Symbols
On the symbol layer, duplicate '*' in a location that exists on a
5-column keyboard. Make room by shifting '@' to the previously empty
spot on the index finger. In this process, move '.' to the ring finger
to facilitate typing relative paths such as '../'.

On the number layer, swap '=' and '.' so that the latter can be reached
more comfortably (by the pinky) while typing sequences of numbers.
Finally, match the location of '%' with the symbol layer and replace its
old location with '~' (for approximate values).