zmk: Reorganize combo keys
Move the problematic Enter combo from OU to .- which is less prone to
accidental rolls while remaining intuitive and ergonomic.
This also eliminates the need for require-prior-idle-ms, which was
frequently forcing me to pause for a short time before pressing Enter at
command prompts.
See commit 842c69dccde840a6c654697a43c6027dcee6f032 for context.

In the process, move Del to XM, and move Space/Backspace to the upper
row on the Lower/Upper layers.