zmk: Improve ergonomics of Symbols layer
Move ` close to " for the `quote Vim mark.
Replace its old position next to . with ~, which is intuitive for typing
Unix file paths.

Move & close to * for references and dereferences.
Relocate _ to the left outer key to make room for this change and
because it is already accessible on the base layer.