zmk: Reclaim outer right key for Enter
Reverts commit 8329de139023d0cf972f729a8274bd31f7298b27

The / and ? symbols are easily reachable on the symbols layer and don't
need to stay on the base layer.
Enter complements Esc on the opposite side nicely in this location.

The .- combo (bottom row) feels alright for Enter but is not as
ergonomic as OU (top row), which I had previously decided to not use for
such a frequently accessed key.